Sync Functionality Check

+++ src/Mirage.Core/Mirage.SocketLayer/ConnectionTrackers/ConnectingTracker.cs
@@ -4,7 +4,7 @@
     {
         private readonly Config _config;
         private readonly Time _time;
-        private float _lastAttempt = float.MinValue;
+        private double _lastAttempt = double.MinValue;
         private int _attemptCount = 0;
 
         public ConnectingTracker(Config config, Time time)
